MLITE-618
MLITE-618

MLITE-618 make sure cash allocation use tenantTime

MLITE-618 implement is active, on cashier

work in progress

implementation of transaction & balance for teller

MLITE-618 wip

MLITE-618 running balance on teller transaction

MLITE_618 adjust journal entrie, add posibility to pull out journal linked to teller

MLITE-618 change migration file name

+++ mifosng-provider/src/main/java/org/mifosplatform/organisation/teller/service/TellerWritePlatformServiceJpaImpl.java
@@ -5,9 +5,14 @@
  */
 package org.mifosplatform.organisation.teller.service;
 
+import java.math.BigDecimal;
+import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import org.joda.time.LocalDate;
+import org.joda.time.LocalDateTime;
 import org.mifosplatform.accounting.common.AccountingConstants.FINANCIAL_ACTIVITY;
 import org.mifosplatform.accounting.financialactivityaccount.domain.FinancialActivityAccount;
 import org.mifosplatform.accounting.financialactivityaccount.domain.FinancialActivityAccountRepositoryWrapper;
@@ -27,6 +32,7 @@ import org.mifosplatform.organisation.office.exception.OfficeNotFoundException;
 import org.mifosplatform.organisation.staff.domain.Staff;
 import org.mifosplatform.organisation.staff.domain.StaffRepository;
 import org.mifosplatform.organisation.staff.exception.StaffNotFoundException;
+import org.mifosplatform.organisation.teller.data.TellerData;
 import org.mifosplatform.organisation.teller.domain.Cashier;
 import org.mifosplatform.organisation.teller.domain.CashierRepository;
 import org.mifosplatform.organisation.teller.domain.CashierTransaction;
@@ -35,9 +41,7 @@ import org.mifosplatform.organisation.teller.domain.CashierTxnType;
 import org.mifosplatform.organisation.teller.domain.Teller;
 import org.mifosplatform.organisation.teller.domain.TellerRepository;
 import org.mifosplatform.organisation.teller.domain.TellerRepositoryWrapper;
-import org.mifosplatform.organisation.teller.exception.CashierExistForTellerException;
-import org.mifosplatform.organisation.teller.exception.CashierNotFoundException;
-import org.mifosplatform.organisation.teller.exception.TellerNotFoundException;
+import org.mifosplatform.organisation.teller.exception.*;
 import org.mifosplatform.organisation.teller.serialization.TellerCommandFromApiJsonDeserializer;
 import org.mifosplatform.portfolio.client.domain.ClientTransaction;
 import org.mifosplatform.useradministration.domain.AppUser;
@@ -63,6 +67,7 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
     private final CashierTransactionRepository cashierTxnRepository;
     private final JournalEntryRepository glJournalEntryRepository;
     private final FinancialActivityAccountRepositoryWrapper financialActivityAccountRepositoryWrapper;
+    private final TellerManagementReadPlatformService tellerManagementReadPlatformService;
 
     @Autowired
     public TellerWritePlatformServiceJpaImpl(final PlatformSecurityContext context,
@@ -70,7 +75,8 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
             final TellerRepositoryWrapper tellerRepositoryWrapper, final OfficeRepository officeRepository,
             final StaffRepository staffRepository, CashierRepository cashierRepository, CashierTransactionRepository cashierTxnRepository,
             JournalEntryRepository glJournalEntryRepository,
-            FinancialActivityAccountRepositoryWrapper financialActivityAccountRepositoryWrapper) {
+            FinancialActivityAccountRepositoryWrapper financialActivityAccountRepositoryWrapper,
+                                             final TellerManagementReadPlatformService tellerManagementReadPlatformService) {
         this.context = context;
         this.fromApiJsonDeserializer = fromApiJsonDeserializer;
         this.tellerRepository = tellerRepository;
@@ -81,6 +87,7 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
         this.cashierTxnRepository = cashierTxnRepository;
         this.glJournalEntryRepository = glJournalEntryRepository;
         this.financialActivityAccountRepositoryWrapper = financialActivityAccountRepositoryWrapper;
+        this.tellerManagementReadPlatformService = tellerManagementReadPlatformService;
     }
 
     @Override
@@ -213,10 +220,9 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
             this.context.authenticatedUser();
             Long hourStartTime;
             Long minStartTime;
-            Long hourEndTime;
-            Long minEndTime;
             String startTime = " ";
-            String endTime = " ";
+            final String endTime=" ";
+
             final Teller teller = this.tellerRepository.findOne(tellerId);
             if (teller == null) { throw new TellerNotFoundException(tellerId); }
             final Office tellerOffice = teller.getOffice();
@@ -225,9 +231,33 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
 
             this.fromApiJsonDeserializer.validateForAllocateCashier(command.json());
 
+
             final Staff staff = this.staffRepository.findOne(staffId);
             if (staff == null) { throw new StaffNotFoundException(staffId); }
+
+
+            // make sure this teller does not yet have an active cashier
+
+            List<Cashier> activeTellerCashier = this.cashierRepository.getActiveTellerCashier(tellerId);
+
+            if(!activeTellerCashier.isEmpty()){
+
+                throw new NoMoreThanOneActiveCashierPerTellerException(tellerId);
+            }
+
+
+            // make sure this cashier is not yet active in another teller
+
+            List<Cashier> activeCashiers = this.cashierRepository.getActiveCashier(staffId, tellerId);
+
+            if(!activeCashiers.isEmpty()){
+
+                throw new CashierAlreadyActiveInAnotherTellerException(staffId);
+            }
+
+
             final Boolean isFullDay = command.booleanObjectValueOfParameterNamed("isFullDay");
+
             if (!isFullDay) {
                 hourStartTime = command.longValueOfParameterNamed("hourStartTime");
                 minStartTime = command.longValueOfParameterNamed("minStartTime");
@@ -237,16 +267,12 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
                 else
                     startTime = hourStartTime.toString() + ":" + minStartTime.toString();
 
-                hourEndTime = command.longValueOfParameterNamed("hourEndTime");
-                minEndTime = command.longValueOfParameterNamed("minEndTime");
-                if (minEndTime == 0)
-                    endTime = hourEndTime.toString() + ":" + minEndTime.toString() + "0";
-                else
-                    endTime = hourEndTime.toString() + ":" + minEndTime.toString();
 
             }
 
-            final Cashier cashier = Cashier.fromJson(tellerOffice, teller, staff, startTime, endTime, command);
+            final Cashier cashier = Cashier.fromJson(tellerOffice, teller, staff, startTime,endTime, command);
+
+            cashier.assign();
 
             this.cashierRepository.save(cashier);
 
@@ -305,6 +331,89 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
 
         return cashierToReturn;
     }
+    @Override
+    @Transactional
+    public CommandProcessingResult assignCashierToTeller(Long tellerId, Long cashierId,JsonCommand command){
+
+        try {
+
+            final AppUser currentUser = this.context.authenticatedUser();
+
+            // make sure this teller does not yet have an active cashier
+
+            List<Cashier> activeTellerCashier = this.cashierRepository.getActiveTellerCashier(tellerId);
+
+            if(!activeTellerCashier.isEmpty()){
+
+               throw new NoMoreThanOneActiveCashierPerTellerException(tellerId);
+            }
+
+
+            // make sure this cashier is not yet active in another teller
+
+            List<Cashier> activeCashiers = this.cashierRepository.getActiveCashier(cashierId, tellerId);
+
+            if(!activeCashiers.isEmpty()){
+
+                throw new CashierAlreadyActiveInAnotherTellerException(cashierId);
+            }
+
+
+            final Cashier cashier = validateUserPriviledgeOnCashierAndRetrieve(currentUser, tellerId, cashierId);
+
+            cashier.assign();
+
+            final Map<String, Object> changes = new LinkedHashMap<>(7);
+            changes.put("active", true);
+
+            this.cashierRepository.saveAndFlush(cashier);
+
+
+            return new CommandProcessingResultBuilder() //
+                    .withCommandId(command.commandId()) //
+                    .withEntityId(cashier.getTeller().getId()) //
+                    .withSubEntityId(cashier.getId()) //
+                    .with(changes) //
+                    .build();
+        } catch (final DataIntegrityViolationException dve) {
+            handleTellerDataIntegrityIssues(command, dve);
+            return CommandProcessingResult.empty();
+        }
+
+    }
+
+    @Override
+    @Transactional
+    public CommandProcessingResult unassignCashierToTeller(Long tellerId, Long cashierId,JsonCommand command){
+
+        try {
+
+            final AppUser currentUser = this.context.authenticatedUser();
+
+            this.fromApiJsonDeserializer.validateForUnassignCashier(command.json());
+
+            final Cashier cashier = validateUserPriviledgeOnCashierAndRetrieve(currentUser, tellerId, cashierId);
+
+            cashier.unassign();
+
+            final Map<String, Object> changes = new LinkedHashMap<>(7);
+            changes.put("active",false);
+
+            this.cashierRepository.saveAndFlush(cashier);
+
+
+            return new CommandProcessingResultBuilder() //
+                    .withCommandId(command.commandId()) //
+                    .withEntityId(cashier.getTeller().getId()) //
+                    .withSubEntityId(cashier.getId()) //
+                    .with(changes) //
+                    .build();
+        } catch (final DataIntegrityViolationException dve) {
+            handleTellerDataIntegrityIssues(command, dve);
+            return CommandProcessingResult.empty();
+        }
+
+    }
 
     @Override
     @Transactional
@@ -312,6 +421,12 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
         try {
             final AppUser currentUser = this.context.authenticatedUser();
             final Cashier cashier = validateUserPriviledgeOnCashierAndRetrieve(currentUser, tellerId, cashierId);
+
+            if(this.tellerManagementReadPlatformService.hasTransaction(cashierId)){
+
+                throw new CashierHasTransactionTellerException(cashierId);
+            }
+
             this.cashierRepository.delete(cashier);
 
         } catch (final DataIntegrityViolationException dve) {
@@ -356,7 +471,18 @@ public class TellerWritePlatformServiceJpaImpl implements TellerWritePlatformSer
             final Cashier cashier = this.cashierRepository.findOne(cashierId);
             if (cashier == null) { throw new CashierNotFoundException(cashierId); }
 
-            this.fromApiJsonDeserializer.validateForCashTxnForCashier(command.json());
+
+            if(txnType.equals(CashierTxnType.SETTLE)){
+
+                final TellerData tellerData = this.tellerManagementReadPlatformService.findTeller(cashier.getTeller().getId());
+
+                this.fromApiJsonDeserializer.validateForCashSettleTxnForCashier(command.json(), tellerData.getBalance());
+            }else{
+                this.fromApiJsonDeserializer.validateForCashTxnForCashier(command.json());
+            }
+
+
+
 
             final String entityType = command.stringValueOfParameterNamed("entityType");
             final Long entityId = command.longValueOfParameterNamed("entityId");
